+++ tools/profiler/public/ProfilerMarkerTypes.h
@@ -37,7 +37,6 @@ namespace geckoprofiler::markers {
 
 // Import some common markers from mozilla::baseprofiler::markers.
 using Tracing = mozilla::baseprofiler::markers::Tracing;
-using FileIO = mozilla::baseprofiler::markers::FileIO;
 using UserTimingMark = mozilla::baseprofiler::markers::UserTimingMark;
 using UserTimingMeasure = mozilla::baseprofiler::markers::UserTimingMeasure;
 using Hang = mozilla::baseprofiler::markers::Hang;
